Replacing scala-reflect with semanticdb-scalac (by chainsaw)
